Language and structure changes
Made a change to the navigation bar. Changed the language file so now it reads Login / Get Started.
Consolidated navigation list items so that the Login and Get Started links share the same <li> element.
Added negative margins on each of the anchor tags containing the modal links and placed a forward slash between them to give similarity between functionality.